+++ configs/allscaip/omol_4m.yml
@@ -134,6 +134,6 @@ runner:
   callbacks:
     - _target_: fairchem.core.common.profiler_utils.ProfilerCallback
       job_config: ${job}
-    - _target_: fairchem.core.components.train.callbacks.TrainCheckpointCallback
+    - _target_: fairchem.core.components.callbacks.TrainCheckpointCallback
       checkpoint_every_n_steps: 2000
       max_saved_checkpoints: 5
