+++ configs/uma/training_release/uma_sm_direct_pretrain.yaml
@@ -174,6 +174,6 @@ runner:
   callbacks:
     - _target_: fairchem.core.common.profiler_utils.ProfilerCallback
       job_config: ${job}
-    - _target_: fairchem.core.components.train.callbacks.TrainCheckpointCallback
+    - _target_: fairchem.core.components.callbacks.TrainCheckpointCallback
       checkpoint_every_n_steps: 5000
       max_saved_checkpoints: 5
