+++ tests/core/units/mlip_unit/test_mlip_train_conserving.yaml
@@ -64,7 +64,7 @@ runner:
   max_steps: ${max_steps}
   max_epochs: ${max_epochs}
   callbacks:
-    - _target_: fairchem.core.components.train.callbacks.TrainCheckpointCallback
+    - _target_: fairchem.core.components.callbacks.TrainCheckpointCallback
       checkpoint_every_n_steps: ${checkpoint_every}
       max_saved_checkpoints: 10
     - _target_: tests.core.units.mlip_unit.test_mlip_unit.TrainEndCallback
